cleanup parity/signer
